requirements.txt
pillow
streamlit
requests
streamlit_lottie
